backend/requirements.txt
# Core web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
python-dotenv>=1.0.0

# AI/ML
anthropic>=0.18.1
numpy>=1.24.0
scipy>=1.10.0

# Voice analysis
resemblyzer>=0.1.3
pydub>=0.25.1
SpeechRecognition>=3.10.1
librosa>=0.10.0

# Face analysis (using OpenCV only for Python 3.14 compatibility)
opencv-python>=4.8.0

# Optional: Install deepface if using Python < 3.13
# deepface>=0.0.80
